fix:修改广告接口bug

+++ business/member-api/src/main/java/com/lawu/eshop/member/api/controller/BankAccountController.java
@@ -20,6 +20,7 @@ import com.lawu.eshop.framework.web.ResultCode;
 import com.lawu.eshop.framework.web.constants.UserConstant;
 import com.lawu.eshop.framework.web.doc.annotation.Audit;
 import com.lawu.eshop.member.api.service.BankAccountService;
+import com.lawu.eshop.member.api.service.CashManageFrontService;
 import com.lawu.eshop.member.api.service.PropertyInfoService;
 import com.lawu.eshop.property.dto.BankAccountDTO;
 import com.lawu.eshop.property.param.BankAccountParam;
@@ -44,6 +45,9 @@ public class BankAccountController extends BaseController{
 	
 	@Autowired
 	private PropertyInfoService propertyInfoService;
+	
+	@Autowired
+	private CashManageFrontService cashManageFrontService;
 
 	@Audit(date = "2017-04-01", reviewer = "孙林青")
 	@Authorization
@@ -76,12 +80,20 @@ public class BankAccountController extends BaseController{
 
     @Audit(date = "2017-04-12", reviewer = "孙林青")
 	@Authorization
-    @ApiOperation(value = "删除银行卡", notes = "删除银行卡[1002]（张荣成）", httpMethod = "DELETE")
+    @ApiOperation(value = "删除银行卡", notes = "删除银行卡[1002|2011]（张荣成）", httpMethod = "DELETE")
     @ApiResponse(code = HttpCode.SC_NO_CONTENT, message = "success")
     @RequestMapping(value = "remove/{id}", method = RequestMethod.DELETE)
     public Result remove(@RequestHeader(UserConstant.REQ_HEADER_TOKEN) String token,
                          @PathVariable @ApiParam(required = true, value = "id") Long id) {
-        Result rs = bankAccountService.delete(id);
+    	String userNum = UserUtil.getCurrentUserNum(getRequest());
+    	Result<Boolean> bankRs= cashManageFrontService.isExistCash(userNum, id);
+    	if(isSuccess(bankRs)){
+    		if(bankRs.getModel()){
+    			return successCreated(ResultCode.BANK_CASH_EXIST);
+    		}else{
+    			 Result rs = bankAccountService.delete(id);
+    		}
+    	}
         return successDelete();
     }
